requirements.txt
python-telegram-bot==21.6
aiosqlite==0.20.0
openai-whisper==20231117
python-dotenv==1.0.1
python-dateutil==2.9.0
requests==2.32.3
anthropic==0.49.0
